pyproject.toml
[project]
authors = [{name = "ryanlinjui", email = "ryanlinjui@gmail.com"}]
dependencies = [
    "gradio>=5.26.0,<6",
    "google-genai>=1.12.1,<2",
    "openai>=1.76.2,<2",
    "pillow>=11.2.1,<12"
, "dotenv>=0.9.9,<0.10"]
description = "Extract structured menu information from images into JSON using a fine-tuned Donut E2E model."
name = "menu-text-detection"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.dependencies]
transformers = ">=4.51.3,<5"
notebook = ">=7.4.0,<8"
datasets = ">=3.5.0,<4"
huggingface_hub = ">=0.30.2,<0.31"
pytorch = ">=2.6.0,<3"
protobuf = ">=5.29.3,<6"
sentencepiece = ">=0.2.0,<0.3"
accelerate = ">=1.5.2,<2"
